✨ Feat: 대시보드 모달 리팩토링 및 식단 수정 버그 픽스 (#148)
